refactor structure auth page
