+++ src/pages/auth/constants.ts
@@ -1,0 +1,1 @@
+export const EMAIL_PATTERN = /[a-z0-9._%+-]+@[a-z0-9.-]+\.[a-z]{2,}$/;
