+++ src/pages/auth/index.ts
@@ -1,2 +1,1 @@
-export * from "./ui";
-export * as authModel from "./model";
+export * from "./view";
