+++ src/pages/auth/view.tsx
